requirements.txt
Django==5.1.2
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
gunicorn==21.2.0
whitenoise==6.6.0